Environmental changes: PHPunit 3.6.x version in now officially supported only. Tests' configuration file has been renamed from browsers.yml(.dist) to config.yml(.dist)
Features added:
1.     1000 new tests
In the new release we introduce complete coverage of Magento CE functionality according to Magento official Smoke Test Suite. This suite covers all test cases that are running during Magento nightly builds testing and we would like the community to use them and expand on them.
2.     Dual screenshots
Upon each error of test failure, a screenshot of currently open pages is captured and stored in .png and html formats at tmp/screenshot folder. In cases were javascript errors are encountered, PNG format is used to capture it. Markup errors are captured in html format.  For easy detection of the screenshot file, the failed test name contacting the class name and timestamp is used. Message containing full path to screenshot is stored in log next to the error message so it can be accessed by direct path.
3.     Multiple UI areas
Magento TAF allows testing of unlimited number of UI areas. The UI areas can be determining in local.yml file for each url to be used by TAF. Based on user-defined urls, TAF automatically detects what set of UI maps should be used and loads the relevant descriptors for current page. With this enhancement, the tester doesn’t need to worry about switching between UI areas within test body. Users can declare separate UImaps for non-magento side services that are specific for your store and write tests for it in the same way written for Magento. This feature is particularly useful when testing several store views or stores using different designs. Each store view or store can have unique UI map set but operate with original set of tests. This feature removes nesting and naming restrictions for UI maps folders and allows the user to organize them as he desire.
Below is a sample of uimap areas declaration:
areas:
                admin:
                    url: 'http://www.localhost.com/magento/admin/'
                    uimap_path: 'admin'
                frontend:
                    url: 'http://www.localhost.com/magento/'
                    uimap_path: 'frontend'
                admin:
                    url: 'http://www.localhost.com/magento/admin/'
                    uimap_path: 'admin'
                paypal:
                    url: ' https://www.paypal.com/ '
                    uimap_path: paypal_ui
                ebay:
                    url: ' http://www.ebay.com/ '
                    uimap_path: ebay